fix: replace double adapter by BigDecimal

+++ src/main/java/org/example/module3/adapter/UsdCurrency.java
@@ -1,5 +1,7 @@
 package org.example.module3.adapter;
 
+import java.math.BigDecimal;
+
 public interface UsdCurrency {
-    double getAmountInUSD();
+    BigDecimal getAmountInUSD();
 }
